# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RunMatch API - runner dating and running partner finder
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runners:
    get:
      operationId: listRunners
      tags: [runners]
      summary: List runners with optional filters
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [date, buddy, both]
          description: Filter by section (date, buddy, or both)
        - name: country
          in: query
          schema:
            type: string
          description: Filter by country
        - name: city
          in: query
          schema:
            type: string
          description: Filter by city
        - name: minPace
          in: query
          schema:
            type: string
          description: Minimum pace (mm:ss per km)
        - name: maxPace
          in: query
          schema:
            type: string
          description: Maximum pace (mm:ss per km)
        - name: experience
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced, elite]
      responses:
        "200":
          description: List of runners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Runner"

    post:
      operationId: createRunner
      tags: [runners]
      summary: Create a new runner profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunnerBody"
      responses:
        "201":
          description: Created runner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
  /runners/{id}:
    get:
      operationId: getRunner
      tags: [runners]
      summary: Get a runner profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Runner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
        "404":
          description: Not found

    put:
      operationId: updateRunner
      tags: [runners]
      summary: Update a runner profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRunnerBody"
      responses:
        "200":
          description: Updated runner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
...
components:
  schemas:
...
    TrackingApps:
      type: object
      properties:
        stravaUrl:
          type: string
          nullable: true
        garminUrl:
          type: string
          nullable: true
        nikeRunClubUrl:
          type: string
          nullable: true
        wahooPlan:
          type: string
          nullable: true
        polarUrl:
          type: string
          nullable: true
        suuntoUrl:
          type: string
          nullable: true
        appleHealthConnected:
          type: boolean
          nullable: true
        garminConnectUrl:
          type: string
          nullable: true
    RunningStats:
      type: object
      properties:
        weeklyMileageKm:
          type: number
          nullable: true
        totalRaces:
          type: integer
          nullable: true
        personalBest5k:
          type: string
          nullable: true
          description: Time in mm:ss format
        personalBest10k:
          type: string
          nullable: true
        personalBestHalfMarathon:
          type: string
          nullable: true
        personalBestMarathon:
          type: string
          nullable: true
        avgPacePerKm:
          type: string
          nullable: true
          description: Pace in mm:ss format
        preferredRunTypes:
          type: array
          items:
            type: string
          nullable: true
          description: e.g. road, trail, track, ultra
    Runner:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        profileType:
          type: string
          enum: [individual, social_club, official_club]
        clubName:
          type: string
          nullable: true
        clubDescription:
          type: string
          nullable: true
        clubWebsite:
          type: string
          nullable: true
        clubSocialUrl:
          type: string
          nullable: true
        clubAssociation:
          type: string
          nullable: true
        lookingFor:
          type: string
          enum: [date, buddy, both]
        experience:
          type: string
          enum: [beginner, intermediate, advanced, elite]
          nullable: true
        trackingApps:
          $ref: "#/components/schemas/TrackingApps"
          nullable: true
        runningStats:
          $ref: "#/components/schemas/RunningStats"
          nullable: true
        lat:
          type: number
          nullable: true
          description: Latitude coordinate
        lng:
          type: number
          nullable: true
          description: Longitude coordinate
        travelCity:
          type: string
          nullable: true
          description: City the runner is currently visiting
        travelCountry:
          type: string
          nullable: true
          description: Country the runner is currently visiting
        travelUntil:
          type: string
          nullable: true
          description: Date until which the runner is travelling (e.g. "May 24")
        travelNote:
          type: string
          nullable: true
          description: Short note about the trip
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - profileType
        - lookingFor
        - createdAt
        - updatedAt
    CreateRunnerBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        profileType:
          type: string
          enum: [individual, social_club, official_club]
        clubName:
          type: string
          nullable: true
        clubDescription:
          type: string
          nullable: true
        clubWebsite:
          type: string
          nullable: true
        clubSocialUrl:
          type: string
          nullable: true
        clubAssociation:
          type: string
          nullable: true
        lookingFor:
          type: string
          enum: [date, buddy, both]
        experience:
          type: string
          enum: [beginner, intermediate, advanced, elite]
          nullable: true
        trackingApps:
          allOf:
            - $ref: "#/components/schemas/TrackingApps"
          minProperties: 1
        runningStats:
          $ref: "#/components/schemas/RunningStats"
          nullable: true
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        travelCity:
          type: string
          nullable: true
        travelCountry:
          type: string
          nullable: true
        travelUntil:
          type: string
          nullable: true
        travelNote:
          type: string
          nullable: true
      required:
        - name
        - profileType
        - lookingFor
        - trackingApps
    UpdateRunnerBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        profileType:
          type: string
          enum: [individual, social_club, official_club]
        clubName:
          type: string
          nullable: true
        clubDescription:
          type: string
          nullable: true
        clubWebsite:
          type: string
          nullable: true
        clubSocialUrl:
          type: string
          nullable: true
        clubAssociation:
          type: string
          nullable: true
        lookingFor:
          type: string
          enum: [date, buddy, both]
        experience:
          type: string
          enum: [beginner, intermediate, advanced, elite]
          nullable: true
        trackingApps:
          $ref: "#/components/schemas/TrackingApps"
          nullable: true
        runningStats:
          $ref: "#/components/schemas/RunningStats"
          nullable: true
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        travelCity:
          type: string
          nullable: true
        travelCountry:
          type: string
          nullable: true
        travelUntil:
          type: string
          nullable: true
        travelNote:
          type: string
          nullable: true
...