# lib/api-zod/src/generated/types/createRunnerBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RunMatch API - runner dating and running partner finder
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRunnerBodyExperience } from './createRunnerBodyExperience';
import type { CreateRunnerBodyLookingFor } from './createRunnerBodyLookingFor';
import type { CreateRunnerBodyProfileType } from './createRunnerBodyProfileType';
import type { RunningStats } from './runningStats';
import type { TrackingApps } from './trackingApps';

export interface CreateRunnerBody {
  name: string;
  age?: number | null;
  bio?: string | null;
  avatarUrl?: string | null;
  city?: string | null;
  country?: string | null;
  gender?: string | null;
  profileType: CreateRunnerBodyProfileType;
  clubName?: string | null;
  clubDescription?: string | null;
  clubWebsite?: string | null;
  clubSocialUrl?: string | null;
  clubAssociation?: string | null;
  lookingFor: CreateRunnerBodyLookingFor;
  experience?: CreateRunnerBodyExperience;
  trackingApps: TrackingApps;
  runningStats?: RunningStats | null;
  lat?: number | null;
  lng?: number | null;
  travelCity?: string | null;
  travelCountry?: string | null;
  travelUntil?: string | null;
  travelNote?: string | null;
}