# lib/api-zod/src/generated/types/updateRunnerBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RunMatch API - runner dating and running partner finder
 * OpenAPI spec version: 0.1.0
 */
import type { RunningStats } from './runningStats';
import type { TrackingApps } from './trackingApps';
import type { UpdateRunnerBodyExperience } from './updateRunnerBodyExperience';
import type { UpdateRunnerBodyLookingFor } from './updateRunnerBodyLookingFor';
import type { UpdateRunnerBodyProfileType } from './updateRunnerBodyProfileType';

export interface UpdateRunnerBody {
  name?: string;
  age?: number | null;
  bio?: string | null;
  avatarUrl?: string | null;
  city?: string | null;
  country?: string | null;
  gender?: string | null;
  profileType?: UpdateRunnerBodyProfileType;
  clubName?: string | null;
  clubDescription?: string | null;
  clubWebsite?: string | null;
  clubSocialUrl?: string | null;
  clubAssociation?: string | null;
  lookingFor?: UpdateRunnerBodyLookingFor;
  experience?: UpdateRunnerBodyExperience;
  trackingApps?: TrackingApps | null;
  runningStats?: RunningStats | null;
  lat?: number | null;
  lng?: number | null;
  travelCity?: string | null;
  travelCountry?: string | null;
  travelUntil?: string | null;
  travelNote?: string | null;
}